feat: remove more unused code
