+++ packages/cli/src/cli/index.ts
@@ -7,7 +7,6 @@ import { CommandBundleAssets } from './config/action.bundle.assets.js';
 import { CommandBundle } from './config/action.bundle.js';
 import { CommandImageryConfig } from './config/action.imagery.config.js';
 import { CommandImport } from './config/action.import.js';
-import { CommandCreateOverview } from './overview/action.create.overview.js';
 import { CommandServe } from './server/action.serve.js';
 
 export class BasemapsConfigCommandLine extends BaseCommandLine {
@@ -22,9 +21,6 @@ export class BasemapsConfigCommandLine extends BaseCommandLine {
     this.addAction(new CommandBundleAssets());
     this.addAction(new CommandImport());
     this.addAction(new CommandImageryConfig());
-
-    // Argo
-    this.addAction(new CommandCreateOverview());
 
     // CICD - Screenshot tests
     this.addAction(new CommandServe());
