+++ packages/server/src/route.layers.ts
@@ -1,46 +1,8 @@
-import { BasemapsConfigProvider, ConfigImagery, getAllImagery, standardizeLayerName } from '@basemaps/config';
-import { Epsg, GoogleTms, Projection, TileMatrixSets } from '@basemaps/geo';
-import { V } from '@basemaps/shared';
+import { BasemapsConfigProvider, getAllImagery } from '@basemaps/config';
+import { Epsg, GoogleTms, TileMatrixSets } from '@basemaps/geo';
+import { getPreviewUrl, V } from '@basemaps/shared';
 
 const previewSize = { width: 1200, height: 630 };
-
-// TODO this should be shared with all of the preview generation logic
-function getImageryCenterZoom(im: ConfigImagery): { lat: number; lon: number; zoom: number } {
-  const center = { x: im.bounds.x + im.bounds.width / 2, y: im.bounds.y + im.bounds.height / 2 };
-
-  // Find a approximate GSD needed to show most of the imagery
-  const aspectWidth = im.bounds.width / previewSize.width;
-  const aspectHeight = im.bounds.height / previewSize.height;
-  const bestAspect = Math.min(aspectHeight, aspectWidth);
-
-  const tms = TileMatrixSets.find(im.tileMatrix);
-  if (tms == null) throw new Error(`Failed to lookup tileMatrix: ${im.tileMatrix}`);
-  const proj = Projection.get(tms);
-  const centerLatLon = proj.toWgs84([center.x, center.y]);
-  const targetZoom = Math.max(tms.findBestZoom(bestAspect) - 7, 0);
-  return { lat: centerLatLon[1], lon: centerLatLon[0], zoom: targetZoom };
-}
-
-export function getPreviewUrl(im: ConfigImagery): {
-  url: string;
-  name: string;
-  locationHash: string;
-  location: { lat: number; lon: number; zoom: number };
-} {
-  const location = getImageryCenterZoom(im);
-  const targetZoom = location.zoom;
-  const lat = location.lat.toFixed(7);
-  const lon = location.lon.toFixed(7);
-  const locationHash = `@${lat},${lon},z${location.zoom}`;
-
-  const name = standardizeLayerName(im.name);
-  return {
-    name,
-    location,
-    locationHash,
-    url: `/v1/preview/${name}/${im.tileMatrix}/${targetZoom}/${lon}/${lat}`,
-  };
-}
 
 export async function createLayersHtml(mem: BasemapsConfigProvider): Promise<string> {
   const allLayers = await mem.TileSet.get('ts_all');
@@ -56,7 +18,7 @@ export async function createLayersHtml(mem: BasemapsConfigProvider): Promise<str
   for (const img of layers) {
     let tileMatrix = TileMatrixSets.find(img.tileMatrix);
     if (tileMatrix == null) tileMatrix = GoogleTms;
-    const ret = getPreviewUrl(img);
+    const ret = getPreviewUrl(img, previewSize);
 
     const els = [
       V('div', { class: `layer-header`, style: 'display:flex; justify-content: space-around;' }, [
